server-side/requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Jinja2==3.1.2
mysqlclient==2.1.1
pycodestyle==2.6.0
python-dotenv==1.0.0
requests==2.18.4
SQLAlchemy==2.0.23
Werkzeug==3.0.1
